+++ src/MockGraphLink.js
@@ -1,4 +1,4 @@
-import { ApolloLink, Observable } from 'apollo-link';
+import { ApolloLink, Observable } from "@apollo/client/core";
 import {
   getMainDefinition,
   getFragmentDefinitions,
@@ -8,14 +8,14 @@ import {
   resultKeyNameFromField,
   isInlineFragment,
   argumentsObjectFromField,
-} from 'apollo-utilities';
-import formatPath from './formatPath';
-import defaultOnError from './defaultOnError';
+} from "@apollo/client/utilities";
+import formatPath from "./formatPath";
+import defaultOnError from "./defaultOnError";
 
-/// copied from graphql-anywhere
+// copied from graphql-anywhere
 export function merge(dest, src) {
-  if (src !== null && typeof src === 'object') {
-    Object.keys(src).forEach(key => {
+  if (src !== null && typeof src === "object") {
+    Object.keys(src).forEach((key) => {
       const srcVal = src[key];
       if (!Object.prototype.hasOwnProperty.call(dest, key)) {
         dest[key] = srcVal;
@@ -29,52 +29,33 @@ export function merge(dest, src) {
 class MockGraphLink extends ApolloLink {
   constructor(getMockGraph, opts = {}) {
     super();
-    const {
-      onError = defaultOnError,
-      fragmentIntrospectionQueryResultData,
-      timeoutMs = 100,
-    } = opts;
+    const { onError = defaultOnError, possibleTypes, timeoutMs = 100 } = opts;
     this.getMockGraph = getMockGraph;
     this.onError = onError;
     this.timeoutMs = timeoutMs;
     this.lastQueryId = 0;
     this.queriesInFlight = new Map();
 
-    if (fragmentIntrospectionQueryResultData) {
-      this.fragmentTypeMap = this.parseFragmentIntrospectionResult(
-        fragmentIntrospectionQueryResultData
-      );
+    if (possibleTypes) {
+      this.possibleTypes = possibleTypes;
     }
-  }
-
-  // Modified version of https://github.com/apollographql/apollo-client/blob/b1ca2025d13a9a446be3b7fda2483767d3f61a6e/packages/apollo-cache-inmemory/src/fragmentMatcher.ts#L148
-  parseFragmentIntrospectionResult(introspectionResultData) {
-    const typeMap = {};
-    introspectionResultData.__schema.types.forEach(type => {
-      if (type.kind === 'UNION' || type.kind === 'INTERFACE') {
-        typeMap[type.name] = type.possibleTypes.map(
-          implementingType => implementingType.name
-        );
-      }
-    });
-    return typeMap;
   }
 
   request(operation, forward) {
     const operationDefinitions = operation.query.definitions;
-    const matchedOperationDefinition = operationDefinitions.find(o =>
-      operation.operationName === null
+    const matchedOperationDefinition = operationDefinitions.find((o) =>
+      operation.operationName == null
         ? o.name == null
         : o.name.value === operation.operationName
     );
 
     let rootValue;
-    if (matchedOperationDefinition.operation === 'query') {
+    if (matchedOperationDefinition.operation === "query") {
       rootValue = this.getMockGraph().Query;
       if (!rootValue) {
         throw new Error('Query must be mocked to fulfill a "query" operation');
       }
-    } else if (matchedOperationDefinition.operation === 'mutation') {
+    } else if (matchedOperationDefinition.operation === "mutation") {
       rootValue = this.getMockGraph().Mutation;
       if (!rootValue) {
         throw new Error(
@@ -83,9 +64,7 @@ class MockGraphLink extends ApolloLink {
       }
     } else {
       throw new Error(
-        `Can't yet mock operation type "${
-          matchedOperationDefinition.operation
-        }"`
+        `Can't yet mock operation type "${matchedOperationDefinition.operation}"`
       );
     }
 
@@ -98,7 +77,7 @@ class MockGraphLink extends ApolloLink {
 
     const executeSelectionSet = (selectionSet, rootValue, currentPath) => {
       const result = {};
-      selectionSet.selections.forEach(selection => {
+      selectionSet.selections.forEach((selection) => {
         if (!shouldInclude(selectionSet, operation.variables)) {
           // Skip this entirely
           return;
@@ -144,25 +123,25 @@ class MockGraphLink extends ApolloLink {
           let fragmentType;
           if (
             fragment.typeCondition &&
-            fragment.typeCondition.kind === 'NamedType' &&
-            fragment.typeCondition.name.kind === 'Name'
+            fragment.typeCondition.kind === "NamedType" &&
+            fragment.typeCondition.name.kind === "Name"
           ) {
             fragmentType = fragment.typeCondition.name.value;
           }
           const objectType = rootValue.__typename;
           if (!objectType) {
             errors.push({
-              type: 'noTypename',
+              type: "noTypename",
               path: currentPath,
               fragmentName: fragment.name.value,
             });
           } else if (fragmentType === objectType) {
             executeFragment();
           } else {
-            if (this.fragmentTypeMap) {
+            if (this.possibleTypes) {
               if (
-                this.fragmentTypeMap[fragmentType] &&
-                this.fragmentTypeMap[fragmentType].indexOf(objectType) !== -1
+                this.possibleTypes[fragmentType] &&
+                this.possibleTypes[fragmentType].indexOf(objectType) !== -1
               ) {
                 executeFragment();
               } else {
@@ -170,7 +149,7 @@ class MockGraphLink extends ApolloLink {
               }
             } else {
               errors.push({
-                type: 'unionInterfaceTypes',
+                type: "unionInterfaceTypes",
                 path: currentPath,
                 fragmentName: fragment.name && fragment.name.value,
                 objectType,
@@ -197,11 +176,11 @@ class MockGraphLink extends ApolloLink {
       let result;
       if (mockedValue === undefined) {
         errors.push({
-          type: 'missing',
+          type: "missing",
           path,
         });
         result = null;
-      } else if (typeof mockedValue === 'function') {
+      } else if (typeof mockedValue === "function") {
         let resolvedValue, executionError;
         try {
           resolvedValue = mockedValue(args || {});
@@ -209,11 +188,11 @@ class MockGraphLink extends ApolloLink {
           executionError = err;
         }
         if (executionError) {
-          errors.push({ type: 'resolver', error: executionError, args, path });
+          errors.push({ type: "resolver", error: executionError, args, path });
           result = null;
         } else if (resolvedValue === undefined) {
           errors.push({
-            type: 'fnReturnUndefined',
+            type: "fnReturnUndefined",
             args,
             path,
           });
@@ -224,10 +203,10 @@ class MockGraphLink extends ApolloLink {
       } else if (
         args &&
         Object.keys(args).length &&
-        typeof mockedValue !== 'function'
+        typeof mockedValue !== "function"
       ) {
         errors.push({
-          type: 'fnRequired',
+          type: "fnRequired",
           args,
           path,
         });
@@ -282,28 +261,26 @@ class MockGraphLink extends ApolloLink {
     );
 
     const formattedErrors = errors.length
-      ? errors.map(e => {
+      ? errors.map((e) => {
           let message = e.type;
-          if (e.type === 'missing') {
+          if (e.type === "missing") {
             message = `Field is missing from mock graph`;
-          } else if (e.type === 'fnReturnUndefined') {
+          } else if (e.type === "fnReturnUndefined") {
             message = `Mock resolver returned undefined; did you mean to return null?`;
-          } else if (e.type === 'fnRequired') {
+          } else if (e.type === "fnRequired") {
             message = `This field received args and thus must be mocked as a function.`;
-          } else if (e.type === 'resolver') {
+          } else if (e.type === "resolver") {
             message = `Error from resolver: ${e.error.message}`;
-          } else if (e.type === 'noTypename') {
+          } else if (e.type === "noTypename") {
+            message = `Can't resolve fragment ${e.fragmentName} because __typename is missing from the mock graph`;
+          } else if (e.type === "unionInterfaceTypes") {
             message = `Can't resolve fragment ${
-              e.fragmentName
-            } because __typename is missing from the mock graph`;
-          } else if (e.type === 'unionInterfaceTypes') {
-            message = `Can't resolve fragment ${
-              e.fragmentName ? e.fragmentName + ' ' : ''
+              e.fragmentName ? e.fragmentName + " " : ""
             }because its type (${
               e.fragmentType
             }) might not match the object's type (${
               e.objectType
-            }). In order to handle union and interface types, you must pass \`fragmentIntrospectionQueryResultData\` when creating MockGraphLink. See https://github.com/dallonf/apollo-link-mock-graph/blob/master/README.md for details.`;
+            }). In order to handle union and interface types, you must pass \`possibleTypes\` when creating MockGraphLink. See https://github.com/dallonf/apollo-link-mock-graph/blob/master/README.md for details.`;
           }
           message = `${formatPath(e.path)}: ${message}`;
           return {
@@ -319,7 +296,7 @@ class MockGraphLink extends ApolloLink {
 
     this.lastQueryId += 1;
     const thisQueryId = this.lastQueryId;
-    const promise = new Promise(resolve =>
+    const promise = new Promise((resolve) =>
       setTimeout(resolve, this.timeoutMs)
     ).then(() => {
       this.queriesInFlight.delete(thisQueryId);
@@ -327,8 +304,8 @@ class MockGraphLink extends ApolloLink {
     });
     this.queriesInFlight.set(thisQueryId, promise);
 
-    const observable = new Observable(sub => {
-      promise.then(result => {
+    const observable = new Observable((sub) => {
+      promise.then((result) => {
         sub.next(result);
         sub.complete();
       });
@@ -343,7 +320,7 @@ class MockGraphLink extends ApolloLink {
       return Promise.all(promises).then(() => {
         if (waitToSettle) {
           // wait a tick for any new queries to start
-          return new Promise(resolve => setTimeout(resolve)).then(() =>
+          return new Promise((resolve) => setTimeout(resolve)).then(() =>
             this.waitForQueries()
           );
         }