+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.0.0
+
+* Breaking: Upgrade to Apollo Client 3
+  * `fragmentIntrospectionQueryResultData` has been replaced with `possibleTypes`; see https://www.apollographql.com/docs/react/migrating/apollo-client-3-migration/#cache-improvements and https://www.apollographql.com/docs/react/data/fragments/#defining-possibletypes-manually for more information on the changes to Apollo Client
+
 # 1.2.0
 
 * Feature: new `timeoutMs` option to customize the delay